+++ astropy/nddata/nduncertainty.py
@@ -7,6 +7,7 @@ import abc
 import numpy as np
 
 from ..utils.compat import ignored
+from ..units import Quantity
 
 __all__ = ['MissingDataAssociationException',
            'IncompatibleUncertaintiesException', 'NDUncertainty',
@@ -150,10 +151,14 @@ class StdDevUncertainty(NDUncertainty):
     support_correlated = False
 
     def __init__(self, array=None, copy=True):
+        self._unit = None
         if array is None:
             self.array = None
         elif isinstance(array, StdDevUncertainty):
             self.array = np.array(array.array, copy=copy, subok=True)
+        elif isinstance(array, Quantity):
+            self.array = np.array(array.value, copy=copy, subok=True)
+            self._unit = array.unit
         else:
             self.array = np.array(array, copy=copy, subok=True)